requirements.txt
colorama
dm-tree
hydra-core
jax[cuda12]==0.4.37
flax==0.10.2
neptune
numpy>=2.1
nvidia-cublas-cu12<12.9
nvidia-cuda-cupti-cu12<12.9
nvidia-cuda-nvcc-cu12<12.9
nvidia-cuda-runtime-cu12<12.9
nvidia-cudnn-cu12<9.7.1
nvidia-cufft-cu12<11.4
nvidia-cusolver-cu12<11.7.3
nvidia-cusparse-cu12<12.5.8
omegaconf
optax
orbax-checkpoint==0.11.1
rich
scipy
tensorflow_probability
tqdm
